requirements.txt
pytest
influxdb==2.11.0
pytest-cov
docker-py
requests
jinja2
PyYAML
kafka-python
